dev-requirements.txt
responses>=0.25.2
pyshacl
mock
pytest-ckan

pytest-cov
pytest-pretty
